+++ R/utils.R
@@ -1,0 +1,6 @@
+max_level_length <- function(x) {
+  if (!is.factor(x))
+    return(0L)
+
+  max(0L, nchar(levels(x)), na.rm = TRUE)
+}